Backend/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
click==8.2.2
contourpy==1.3.3
cycler==0.12.1
deprecation==2.1.0
fastapi==0.116.1
flatbuffers==25.2.10
fonttools==4.59.0
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jax==0.7.0
jaxlib==0.7.0
kiwisolver==1.4.8
matplotlib==3.10.5
mediapipe==0.10.21
ml_dtypes==0.5.3
numpy==2.3.2
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
postgrest==1.1.1
protobuf==4.25.8
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
realtime==2.6.0
scipy==1.16.1
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
starlette==0.47.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
websockets==15.0.1
